I a moved some util function into the same file, and i fix bug in the api responese by adding a return statement after each call to the send function.
