+++ src/auth_endpoints/login.auth.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
 
-export async function loginAuth(req: Request, res: Response, next: any) {
+export async function loginAuth(req: Request, res: Response) {
 	res.send("muvees backend login authentication API endpoint called");
 }
