--- src/utils/auth/generate.hash.salt.ts
@@ -1,11 +1,0 @@
-import crypto from "crypto";
-
-export async function generatePasswordSalt(): Promise<string> {
-	const salt = crypto
-		.randomBytes(256)
-		.toString("base64")
-		.replace(/\+/g, "x")
-		.replace(/\s/g, "v");
-
-	return salt;
-}
