+++ tsconfig.json
@@ -3,8 +3,8 @@
 	"exclude": [],
 	"compilerOptions": {
 		"target": "ES2017",
-		"module": "NodeNext",
-		"moduleResolution": "NodeNext",
+		"module": "Node16",
+		"moduleResolution": "Node16",
 		"esModuleInterop": true,
 		"forceConsistentCasingInFileNames": true,
 		"strict": true,
